fix(task): only observe BusinessProcessEvents after transactions have succeeded
taskEvent might be called for tasks, even if a rollback occurs and they don't actually get persisted. This might lead to "ghost tasks" being synchronized to Elastic, which can't be opened in Alanda (because they don't exist).
This gets worse when the failing task is retried (either by Camunda's retry mechanism, or when a OptimisticLockingException occurs), syncing multiple "ghost tasks" to Elastic.

See also:
- https://docs.camunda.org/manual/7.4/user-guide/cdi-java-ee-integration/the-cdi-event-bridge/